#29 make UI for inspection ad

Closes #29

See merge request tdt4140-2023/landsby-3/gruppe-49/toolbox!35

+++ toolbox/src/lib/controller.ts
@@ -1,8 +1,7 @@
-import { addDoc, collection, collectionGroup, deleteDoc, doc, getDoc, getDocs, getFirestore, query, updateDoc, where } from "firebase/firestore";
+import { addDoc, collection, collectionGroup, deleteDoc, doc, DocumentData, getDoc, getDocs, getFirestore, query, updateDoc, where } from "firebase/firestore";
 import { app, storage } from "./firebase";
-import { NewUser } from "../types/types";
+import { Ad, NewUser } from "../types/types";
 import { getDownloadURL, ref, uploadBytes } from "firebase/storage";
-
 
 const firestore = getFirestore(app);
 
@@ -18,49 +17,96 @@ export const addUser = async (userData: NewUser) => {
 
 // CREATE A NEW ADD
 export const addAd = async (adData: any) => {
-  const userId = sessionStorage.getItem("userID"); // Get userId from sessionStorage
-  const newAd = await addDoc(collection(firestore, `users/${userId}/ads`), {...adData});
+  const newAd = await addDoc(collection(firestore, `adsCol`), {...adData});
   console.log(`New ad created${newAd.path}`)
 };
 
 
 // READ
 export const usersCollection = collection(firestore, "users"); // USERS COLLECTION
-export const adsCollection = collectionGroup(firestore, "ads"); // ADDS COLLECTION
+export const adsCollection = collection(firestore, "adsCol"); // ADDS COLLECTION
+
+
 // GET SPECIFIC USER
 export const getUser = async (id: string) => {
   const document = doc(firestore, `users/${id}`);
   const user = await getDoc(document);
-  console.log(`User with ID: ${id} found`);
+  console.log(`getUser: User with ID: ${id}`);
   return user;
 };
+
+// Get ad when clicking on ad
+export const getAd = async (adId: string) => {
+  const document = doc(firestore, `adsCol/${adId}`);
+  const ad = await getDoc(document).then((ad) => {
+  if (ad.exists()) {
+    ad.data().userid !== undefined ? sessionStorage.setItem("userIDFromAd", ad.data().userid) : console.log("No user ID");
+    return ad;
+  }
+  return ad;
+  });
+  return ad;
+};
+
+// Get all ads from a specific user
+export const getUserAds = async (id: string) => {
+  const querySnapshot = await getDocs(collection(firestore, `adsCol`));
+
+  const userAds: DocumentData[] = [];
+  const userAds2: Ad[] = [];
+  const documentID: string[] = [];
+
+  querySnapshot.forEach((doc) => {
+    if (doc.data().userid === id) {
+      // push id for doc
+      documentID.push(doc.id);
+      userAds.push(doc.data());
+    }
+  });
+
+  userAds.forEach((ad, index) => {
+    userAds2.push({
+      id: documentID[index],
+      userid: ad.userid,
+      title: ad.title,
+      description: ad.description,
+      category: ad.category,
+      price: ad.price,
+      rental: ad.rental,
+      address: ad.address,
+      zip: ad.zip,
+      city: ad.city,
+      pictures: ad.pictures,
+    });
+  });
+  return userAds2;
+};
+
+
+
+
+
+
 
 
 // UPDATE 
 // UPDATE A USER
-export const updateUser = async (id: string, userData: NewUser) => {
-  const document = doc(firestore, `users/${id}`);
-  await updateDoc(document, {...userData});
-  console.log(`Updated user with ID: ${id}`);
-};
+
 
 // UPDATE A ADD
-export const updateAd = async (adId: string, adData: any) => {
-  const userId = sessionStorage.getItem("userId"); // Get userId from sessionStorage
-  const document = doc(firestore, `users/${userId}/ads/${adId}`);
-  await updateDoc(document, {...adData});
-  console.log(`Updated ad with ID: ${adId}`);
-}
+
 
 
 
 // DELETE
 // HELP FUNCTION FOR DELETING ALL ADS IN USERS COLLECTION
 export const deleteUserAdsCollection = async (id: string) => {
-    const querySnapshot = await getDocs(collection(firestore, `users/${id}/ads`));
-    querySnapshot.forEach((doc) => {
+  const querySnapshot = await getDocs(collection(firestore, `adsCol`));
+  querySnapshot.forEach((doc) => {
+      if (doc.data().userId === id) {
         deleteDoc(doc.ref);
-    });
+      }
+  });
 };
 
 // DELETE A USER
@@ -75,15 +121,14 @@ export const deleteUser = async (id: string) => {
 
 
 // DELETE A AD
-export const deleteAd = async ( adId: string ) => { 
-    const userID = sessionStorage.getItem("userID"); // Get userId from sessionStorage
-    const adsDoc = doc(firestore, `users/${userID}/ads/${adId}`);
-    if (adsDoc) { //TODO: Fiks denne if-setningen. Må sjekke om doc eksisterer i databasen på denne brukeren.  
+  export const deleteAd = async ( adId: string ) => { 
+    const adsDoc = doc(firestore, `adsCol/${adId}`);
+    if (adsDoc) { //TODO: Fiks denne if-setningen. Må sjekke om doc eksisterer i databasen på denne brukeren.
       await deleteDoc(adsDoc).then(() => {
-        console.log(`Deleted ad with ID: ${adId} from user with ID: ${userID}`);
+        console.log(`Deleted ad with ID: ${adId}`);
       });
     } else {
-      console.log(`No ad with ID: ${adId} found in user with ID: ${userID}`);
+      console.log(`No ad with ID: ${adId} found`);
     }
   };
   
@@ -93,7 +138,7 @@ export const deleteAd = async ( adId: string ) => {
 export function addToSessionStorage(username: string) { // Brukes i loginpage og registerpage
   sessionStorage.setItem("username", username ); 
   console.log("Username sessionstorage set to: " + sessionStorage.getItem("username"));
-  const dummy = getDocs(query(usersCollection, where("username", "==", username))).then((querySnapshot) => {
+  getDocs(query(usersCollection, where("username", "==", username))).then((querySnapshot) => {
     querySnapshot.forEach((doc) => {
       const userID = doc.id; // Itererer gjennom alle brukere i databasen og setter userID lik brukerens ID
       sessionStorage.setItem("userID", userID);
@@ -111,7 +156,7 @@ export function removeFromSessionStorage() { // Brukes i logoutbutton
 
 export async function validateUser(username: string, password: string) { // Brukes i loginpage
   let a = false;
-  const dummy = await getDocs(query(usersCollection, where("username", "==", username))).then((querySnapshot) => {
+  await getDocs(query(usersCollection, where("username", "==", username))).then((querySnapshot) => {
     querySnapshot.forEach((doc) => {
       const userPassword = doc.data().password;
       if (password === userPassword && username === doc.data().username) {
@@ -126,7 +171,7 @@ export async function validateUser(username: string, password: string) { // Bruk
 
 export async function validateDuplicateUsername(username: string) { // Brukes i registerpage
   let a = false;
-  const dummy = await getDocs(query(usersCollection, where("username", "==", username))).then((querySnapshot) => {
+  await getDocs(query(usersCollection, where("username", "==", username))).then((querySnapshot) => {
     querySnapshot.forEach((doc) => {
       if (username === doc.data().username) {
         a = true;
@@ -140,8 +185,6 @@ export async function validateDuplicateUsername(username: string) { // Brukes i 
 
 // upload image to firebase storage
 export const uploadImage = async (file: any) => {
-  console.log("file backend: " + file);
-  console.log("Upload image backend")
   const storageRef = ref(storage, `images/${file.name}`);
   const response = await uploadBytes(storageRef, file);
   const url = await getDownloadURL(response.ref);
@@ -155,7 +198,6 @@ export const uploadImages = async (files: any) => {
     const url = await uploadImage(files[i]);
     urls.push(url);
   }
-  console.log("urls backend " + urls);
   return urls;
 }
